Add tests for the bugfix
